added attachments plugin and created slider inside post

+++ single.php
@@ -1,14 +1,14 @@
 <?php
 $alpha_layout_class = "col-md-8";
-$alpha_text_class = "";
-if ( !is_active_sidebar( "sidebar-1" ) ) {
+$alpha_text_class   = "";
+if ( ! is_active_sidebar( "sidebar-1" ) ) {
     $alpha_layout_class = "col-md-10 offset-md-1";
-    $alpha_text_class = "text-center";
+    $alpha_text_class   = "text-center";
 }
 ?>
 
 <?php get_header(); ?>
-<body <?php body_class(array('first_class','second_class')); ?>>
+<body <?php body_class( array( "first_class", "second_class" ) ); ?>>
 <?php get_template_part( "/template-parts/common/hero" ); ?>
     <div class="container">
         <div class="row">
@@ -18,7 +18,7 @@ if ( !is_active_sidebar( "sidebar-1" ) ) {
                     while ( have_posts() ) :
                         the_post();
                         ?>
-                        <div <?php post_class(array('first_post_class','second_post_class')); ?>>
+                        <div <?php post_class( array( "first_class", "second_class" ) ); ?>>
                             <div class="container">
                                 <div class="row">
                                     <div class="col-md-12">
@@ -26,7 +26,6 @@ if ( !is_active_sidebar( "sidebar-1" ) ) {
                                             <?php the_title(); ?>
                                         </h2>
                                         <p class="<?php echo $alpha_text_class; ?>">
-<!--                                            <em>--><?php //the_author(); ?><!--</em><br/>-->
                                             <em><?php the_author_posts_link(); ?></em><br/>
                                             <?php echo get_the_date(); ?>
                                         </p>
@@ -34,21 +33,37 @@ if ( !is_active_sidebar( "sidebar-1" ) ) {
                                 </div>
                                 <div class="row">
                                     <div class="col-md-12">
-                                        <p>
+                                        <div class="slider" id="customize">
                                             <?php
-                                            if ( has_post_thumbnail() ) {
-                                                $thumbnail_url = get_the_post_thumbnail_url( null, "large" );
-                                                printf( '<a class="popup" href="%s" data-featherlight="image">', $thumbnail_url );
-                                                the_post_thumbnail( "large", array( "class" => "img-fluid" ) );
-                                                echo '</a>';
+                                            if ( class_exists( 'Attachments' ) ) {
+                                                $attachments = new Attachments( 'slider' );
+                                                if ( $attachments->exist() ) {
+                                                    while ( $attachment = $attachments->get() ) { ?>
+                                                        <div>
+                                                            <?php echo $attachments->image( 'large' ); ?>
+                                                        </div>
+                                                        <?php
+                                                    }
+                                                }
+                                            }
+                                            ?>
+                                        </div>
+                                        <div>
+                                            <?php
+                                            if ( !class_exists( 'Attachments' ) ) {
+                                                if ( has_post_thumbnail() ) {
+                                                    $thumbnail_url = get_the_post_thumbnail_url( null, "large" );
+                                                    printf( '<a class="popup" href="%s" data-featherlight="image">', $thumbnail_url );
+                                                    the_post_thumbnail( "large", array( "class" => "img-fluid" ) );
+                                                    echo '</a>';
+                                                }
                                             }
 
                                             the_content();
-
                                             wp_link_pages();
 
                                             ?>
-                                        </p>
+                                        </div>
                                     </div>
 
                                     <div class="authorsection">
